python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "reterminal"
version = "2.0.0"
description = "HTTP client and CLI for Seeed reTerminal E1001 ePaper display"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "amadad"}
]
dependencies = [
    "requests>=2.31.0",
    "Pillow>=10.0.0",
    "typer>=0.9.0",
    "loguru>=0.7.0",
    "tenacity>=8.2.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
qr = [
    "segno>=1.5.0",
]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]
all = [
    "reterminal[qr,dev]",
]

[project.scripts]
reterminal = "reterminal.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["reterminal*"]

[tool.ruff]
line-length = 100
target-version = "py310"
